fix(emails): also repair mojibake in inbox-list previews (sender, subject, snippet)
User pointed out: the email-body iframe shows "We're" correctly now, but
the sidebar list still shows "Du hast 250,00Â âÂ EUR" — because the
mojibake repair lived inside EmailBodyFrame and never touched the
preview/subject/sender strings that go into the list row.

Extract repairMojibake() into a shared src/lib/mojibake.ts utility so:
  - EmailBodyFrame keeps using it for body rendering (no behavior change)
  - useLiveEmails now applies it to preview + sender + subject before
    they flow into the LiveEmail UI shape

Net effect: receipts ("250,00 €"), apostrophes in subjects ("We're"),
and German umlauts in sender names all render correctly in the inbox
list, not just inside an opened mail.

Tests still 7/7 green.
